+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.18.9
+
+* Adds support for `onMapLoaded` callback.
+
 ## 2.18.8
 
 * Bumps com.google.maps.android:android-maps-utils from 3.6.0 to 3.19.1.